+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Changed
 
-- `addons/br`: **breaking**: the Brazil NF-e/NFC-e (`br-nfe-v4`) and NFS-e (`br-nfse-v1`) addons moved to the standalone [`github.com/invopop/gobl.br`](https://github.com/invopop/gobl.br) module. Add a blank import of `gobl.br/addon` to keep using the `br-nfe-*` / `br-nfse-*` addon keys.
+- `addons/br`: **breaking**: the Brazil NF-e/NFC-e (`br-nfe-v4`) and NFS-e (`br-nfse-v1`) addons moved to their own standalone modules, [`github.com/invopop/gobl.br.nfe`](https://github.com/invopop/gobl.br.nfe) and [`github.com/invopop/gobl.br.nfse`](https://github.com/invopop/gobl.br.nfse) respectively. Add a blank import of `gobl.br.nfe/addon` and/or `gobl.br.nfse/addon` to keep using the `br-nfe-*` / `br-nfse-*` addon keys.
 
 ## [v0.501.0] - 2026-06-16